+++ evaluator/values/ArrayValue.hpp
@@ -12,12 +12,12 @@ namespace soviet {
         ArrayValue();
 
         void add(std::shared_ptr<Value> value);
-        inline std::shared_ptr<Value>& at(const unsigned int index);
+        inline std::shared_ptr<Value>& at(const size_t index);
         inline Data& getData();
         size_t size() const;
         bool equals(const std::shared_ptr<Value>& other) override;
         std::shared_ptr<Value> clone() override;
-        void removeAt(unsigned int index);
+        void removeAt(size_t index);
         void remove(const std::shared_ptr<Value>& element);
         void concat(const std::shared_ptr<ArrayValue>& other);
         void reverse();
